GameWindow: Moved askUsername method into Game
